// src/app/lib/generated/models/Occupation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Occupation = {
    Student: 'student',
    Teacher: 'teacher',
    Doctor: 'doctor',
    Soldier: 'soldier',
    Nurse: 'nurse',
    Therapist: 'therapist',
    Model: 'model',
    Policeman: 'policeman',
    Domina: 'domina',
    YogaInstruction: 'yoga_instruction',
    Lawyer: 'lawyer',
    Stewardess: 'stewardess',
    BusinessOwner: 'business_owner',
    Housewife: 'housewife',
    Dj: 'DJ'
} as const;
export type Occupation = typeof Occupation[keyof typeof Occupation];


export function instanceOfOccupation(value: any): boolean {
    for (const key in Occupation) {
        if (Object.prototype.hasOwnProperty.call(Occupation, key)) {
            if (Occupation[key as keyof typeof Occupation] === value) {
                return true;
            }
        }
    }
    return false;
}

export function OccupationFromJSON(json: any): Occupation {
    return OccupationFromJSONTyped(json, false);
}

export function OccupationFromJSONTyped(json: any, ignoreDiscriminator: boolean): Occupation {
    return json as Occupation;
}

export function OccupationToJSON(value?: Occupation | null): any {
    return value as any;
}

export function OccupationToJSONTyped(value: any, ignoreDiscriminator: boolean): Occupation {
    return value as Occupation;
}