// src/app/lib/generated/models/Hobby.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Hobby = {
    Reading: 'reading',
    Writing: 'writing',
    Youtube: 'youtube',
    TvSeries: 'tv_series',
    Anime: 'anime',
    Cosplay: 'cosplay',
    Hiking: 'hiking',
    Manga: 'manga',
    Streaming: 'streaming',
    Yoga: 'yoga',
    TikTok: 'tik_tok',
    Photography: 'photography',
    Traveling: 'traveling',
    Dancing: 'dancing',
    VideoGames: 'video_games',
    Memes: 'memes',
    Podcasts: 'podcasts',
    DigitalArt: 'digital_art',
    Movies: 'movies',
    Fitness: 'fitness',
    Cooking: 'cooking',
    Fishing: 'fishing',
    Camping: 'camping',
    Chess: 'chess',
    BoardGames: 'board_games',
    Vlogging: 'vlogging',
    Sports: 'sports',
    Puzzles: 'puzzles'
} as const;
export type Hobby = typeof Hobby[keyof typeof Hobby];


export function instanceOfHobby(value: any): boolean {
    for (const key in Hobby) {
        if (Object.prototype.hasOwnProperty.call(Hobby, key)) {
            if (Hobby[key as keyof typeof Hobby] === value) {
                return true;
            }
        }
    }
    return false;
}

export function HobbyFromJSON(json: any): Hobby {
    return HobbyFromJSONTyped(json, false);
}

export function HobbyFromJSONTyped(json: any, ignoreDiscriminator: boolean): Hobby {
    return json as Hobby;
}

export function HobbyToJSON(value?: Hobby | null): any {
    return value as any;
}

export function HobbyToJSONTyped(value: any, ignoreDiscriminator: boolean): Hobby {
    return value as Hobby;
}